backend/pyproject.toml
[project]
name = "paperfin-backend"
version = "0.1.0"
description = "Jellyfin-style paper scraper and subscription tool"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.22",
    "pydantic-settings>=2.5.0",
    "python-multipart>=0.0.9",
    "pypdf>=5.0.0",
    "pdfplumber>=0.11.0",
    "pymupdf>=1.24.0",
    "arxiv>=2.1.3",
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    "anthropic>=0.40.0",
    "apscheduler>=3.10.4",
    "Pillow>=10.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
ignore = ["E501"]
